First version of the SMU program.
This is the first commit of the keithley SMU program. Right now it is
in a very rough status and a lot of work needs to be done. Despite this,
this version does work and therefore will count as the first version.

Right now, the configure reverts to default every time you reopen the
dialog. Also, I need to implement a way for changing the sweep the
program performs. I would like to add multithreading and a real-time
graph as well. This is a list of some things I plan to do going forward,
although there are no "changes" this time since it is the first working
version and everything is new.