No config. dialog reset.
Fixed the config. dialog so that it doesn't reset every time you open
it. Started messing around with settings to cause the settings to be
persistant every time you use the program, but have not done much with
that. The changes made to try and do this don't seem to have broken the
program so I will push it now before that happens. This way I have a
working version to use while I tinker with getting the settings to be
persistant.