Persistance compliance level and data path working.
Made it so the compliance level and data path stay the same on program
reload using QSettings. Now I would like to go through and do this for
the other persistant settings.

Doing the other boxes will be quick but it will take a little trickery
to get the pull down boxes to work.